# Speedup of force exchange + force exchange debug mode fix