--- a/examples/force-exchange/cube-dense-nt.xml
+++ b/examples/force-exchange/cube-dense-nt.xml
@@ -1,0 +1,105 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<mardyn version="20100525" >
+
+  <simulation type="MD" >
+    <integrator type="Leapfrog" >
+      <!-- MD ODE integrator -->
+      <timestep unit="reduced" >0.002</timestep>
+    </integrator>
+
+    <run>
+      <currenttime>0</currenttime>
+      <production>
+        <steps>100</steps>
+      </production>
+    </run>
+
+    <ensemble type="NVT">
+      <!-- Nparticles determined by the input file -->
+      <temperature unit="reduced" >0.95</temperature>
+      <domain type="box">
+          <lx>80.0</lx>
+          <ly>80.0</ly>
+          <lz>80.0</lz>
+      </domain>
+
+      <components>
+          <moleculetype id="1" name="Argon">
+            <site type="LJ126" id="1" >
+              <coords> <x>0.0</x> <y>0.0</y> <z>0.0</z> </coords>
+              <mass>1.0</mass>
+              <sigma>1.0</sigma>
+              <epsilon>1.0</epsilon>
+            </site>
+          </moleculetype>
+      </components>
+
+      <phasespacepoint>
+        <generator name="mkTcTS">
+          <layer1>
+            <density>1</density>
+          </layer1>
+          <layer2>
+            <density>1</density>
+          </layer2>
+        </generator>
+      </phasespacepoint>
+    </ensemble>
+
+    <algorithm>
+      <!--
+      <parallelisation type="KDDecomposition">
+        <updateFrequency>10</updateFrequency>
+        <fullSearchThreshold>2</fullSearchThreshold>
+      </parallelisation>
+      -->
+      <parallelisation type="DomainDecomposition">
+      	<CommunicationScheme>indirect</CommunicationScheme>
+      </parallelisation>
+      <datastructure type="LinkedCells">
+      	  <traversalSelector>nt</traversalSelector>
+          <cellsInCutoffRadius>2</cellsInCutoffRadius>
+      </datastructure>
+      <cutoffs type="CenterOfMass" >
+        <radiusLJ unit="reduced" >5.0</radiusLJ>
+      </cutoffs>
+      <electrostatic type="ReactionField" >
+        <epsilon>1.0e+10</epsilon>
+      </electrostatic>
+
+      <!--
+      <thermostats>
+        <thermostat type="VelocityScaling" componentName="Argon">
+          <temperature>1</temperature>
+        </thermostat>
+      </thermostats>
+      -->
+    </algorithm>
+
+    <output>
+    <!--
+      <outputplugin name="SysMonOutput">
+        <expression>procloadavg:loadavg1</expression>
+        <expression label="Free+BufferRAM [MB]">sysinfo:freeram sysinfo:bufferram + sysinfo:mem_unit * float 1048576 /</expression>
+        <expression>procselfstatm:size 1048576. /</expression>
+        <writefrequency>10</writefrequency>
+      </outputplugin>
+      -->
+      <!-- <outputplugin name="Resultwriter">
+        <writefrequency>1</writefrequency>
+        <outputprefix>mkTcTS</outputprefix>
+      </outputplugin> -->
+      <!--
+      <outputplugin name="CheckpointWriter">
+        <writefrequency>10</writefrequency>
+        <outputprefix>mkTcTS</outputprefix>
+      </outputplugin>
+      -->
+      <!--<outputplugin name="PovWriter">
+        <writefrequency>10</writefrequency>
+        <outputprefix>mkTcTS</outputprefix>
+      </outputplugin> -->
+    </output>
+
+  </simulation>
+</mardyn>