--- a/src/particleContainer/ParticleContainer.h
+++ b/src/particleContainer/ParticleContainer.h
@@ -21,11 +21,14 @@
 #define PARTICLECONTAINER_H_
 
 #include <list>
+#include <variant>
 #include <vector>
+#include "ParticleCell.h"
 #include "ParticleIterator.h"
 #include "RegionParticleIterator.h"
-#include "molecules/MoleculeForwardDeclaration.h"
 #include "io/MemoryProfiler.h"
+#include "molecules/MoleculeForwardDeclaration.h"
+
 class CellProcessor;
 class ParticlePairsHandler;
 class XMLfileUnits;
@@ -225,10 +228,9 @@ public:
 	/**
 	 * @brief Gets a molecule by its position.
 	 * @param pos Molecule position
-	 * @param result Molecule will be returned by this pointer if found
-	 * @return Molecule was found?
+	 * @return Iterator to the molecule. The iterator is invalid if no molecule was found.
 	 */
-	virtual bool getMoleculeAtPosition(const double pos[3], Molecule** result) = 0;
+	virtual std::variant<ParticleIterator, SingleCellIterator<ParticleCell>> getMoleculeAtPosition(const double pos[3]) = 0;
 
 	// @brief Should the domain decomposition exchange calculated forces at the boundaries,
 	// or does this particle container calculate all forces.