# Description

Introduces a speedup to the force exchange:
- Instead of searching for every particle within a cell, the iterator is now reused when possible.

BugFix(es):
- Fixes first bytes of force buffer containing `_numForces`, even though they are assumed not to be there. (previously assertions were thrown in debug mode when multiple halo regions exist for one communication partner).

# How Has This Been Tested?

Please describe the tests that you ran to verify your changes.

- [x] Run tests locally to check correctness.

# Performance impact

20k particles

Before:

num_cpu | time_es | time_fs | time_hs
----------- | -------- | -------- | ------
1 |  28.447 | 34.067 | 35.94
2 | 15.894 | 18.864 | 22.127
4 | 8.9616 | 10.257 | 14.189
8 | 5.2629 | 5.6653 | 8.7165

After:
num_cpu | time_es | time_fs | time_hs
----------- | -------- | -------- | ------
1 | 22.82 | 33.678 | 23.267
2 | 11.726 | 18.838 | 12.353
4 | 6.0614 | 10.247 | 6.6284
8 | 3.1293 | 5.5981 | 3.5631

- no change for fs (as expected)
- significant speedup for ES, HS.

Plots:

- use FS as reference (it  remains almost unchanged!)

unoptimized;
![zonal_methods_20k_loop](https://user-images.githubusercontent.com/38119442/104563670-3a437480-564a-11eb-8260-f42ae576426e.png)

optimized;
![zonal_methods_20k_loop](https://user-images.githubusercontent.com/38119442/104563694-43ccdc80-564a-11eb-9982-939179164508.png)